Extract codex-auth from codex-core
