+++ codex-rs/login/Cargo.toml
@@ -11,6 +11,7 @@ workspace = true
 base64 = { workspace = true }
 chrono = { workspace = true, features = ["serde"] }
 codex-client = { workspace = true }
+codex-auth = { workspace = true }
 codex-core = { workspace = true }
 codex-app-server-protocol = { workspace = true }
 rand = { workspace = true }
