+++ codex-rs/login/src/lib.rs
@@ -2,6 +2,13 @@ mod device_code_auth;
 mod pkce;
 mod server;
 
+pub use codex_auth::AuthDotJson;
+pub use codex_auth::CODEX_API_KEY_ENV_VAR;
+pub use codex_auth::OPENAI_API_KEY_ENV_VAR;
+pub use codex_auth::TokenData;
+pub use codex_auth::login_with_api_key;
+pub use codex_auth::logout;
+pub use codex_auth::save_auth;
 pub use codex_client::BuildCustomCaTransportError as BuildLoginHttpClientError;
 pub use device_code_auth::DeviceCode;
 pub use device_code_auth::complete_device_code_login;
@@ -16,11 +23,4 @@ pub use server::run_login_server;
 pub use codex_app_server_protocol::AuthMode;
 pub use codex_core::AuthManager;
 pub use codex_core::CodexAuth;
-pub use codex_core::auth::AuthDotJson;
 pub use codex_core::auth::CLIENT_ID;
-pub use codex_core::auth::CODEX_API_KEY_ENV_VAR;
-pub use codex_core::auth::OPENAI_API_KEY_ENV_VAR;
-pub use codex_core::auth::login_with_api_key;
-pub use codex_core::auth::logout;
-pub use codex_core::auth::save_auth;
-pub use codex_core::token_data::TokenData;
