+++ codex-rs/login/src/server.rs
@@ -28,13 +28,13 @@ use crate::pkce::generate_pkce;
 use base64::Engine;
 use chrono::Utc;
 use codex_app_server_protocol::AuthMode;
+use codex_auth::AuthCredentialsStoreMode;
+use codex_auth::AuthDotJson;
+use codex_auth::TokenData;
+use codex_auth::parse_chatgpt_jwt_claims;
+use codex_auth::save_auth;
 use codex_client::build_reqwest_client_with_custom_ca;
-use codex_core::auth::AuthCredentialsStoreMode;
-use codex_core::auth::AuthDotJson;
-use codex_core::auth::save_auth;
 use codex_core::default_client::originator;
-use codex_core::token_data::TokenData;
-use codex_core::token_data::parse_chatgpt_jwt_claims;
 use rand::RngCore;
 use serde_json::Value as JsonValue;
 use tiny_http::Header;
